pyproject.toml
[tool.poetry]
name = "meld-cds-hook"
version = "0.1.0"
description = ""
authors = ["Christopher Fu <invent288cfu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
requests = "^2.32.3"
aiohttp = "^3.9.5"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^8.3.1"
bump-pydantic = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
